% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chi2test.R
\name{chi2test}
\alias{chi2test}
\title{Create a chi-square correlation matrix of an input data.frame}
\usage{
chi2test(matrix, dim = 1, chi2limit = 0.05)
}
\arguments{
\item{matrix}{data.frame with numeric values}

\item{dim}{switch to define, whether the correlation matrix should be created
for columns or rows.
1 (default): table is created for column (variables) relations.
2: table is created for row (objects) relations.}

\item{chi2limit}{significance level for the test decision. default: 0.05 -> 5\%}
}
\value{
chi-square correlation matrix
}
\description{
\code{chi2test} returns a correlation matrix of a data.frame. The correlation value
is the test decision of the chi-square test for a defined decision niveau. A significant
relation of two variables/objects is marked with a numeric "1", a negativ test result however
with a numeric "0". The correlation matrix can be created for column or row relations of the
input data.frame.
}
\examples{
testmatrixrand <- data.frame(
matrix(base::sample(0:1,400,replace=T), nrow=20, ncol=20)
)

# correlation table is created for the columns of the input data.frame testmatrixrand
chi2test(testmatrixrand)
chi2test(testmatrixrand, 1, 0.1)
chitabcolumns <- chi2test(matrix = testmatrixrand, dim = 1, chi2limit = 0.1)

# correlation table is created for the rows of the input data.frame testmatrixrand
chi2test(testmatrixrand, 2)
chitabrows <- chi2test(matrix = testmatrixrand, dim = 2, chi2limit = 0.03)
}

