% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/corrmat.R
\name{corrmat}
\alias{corrmat}
\title{Create a correlation matrix of an input data.frame}
\usage{
corrmat(matrix, method = "chi2", dim = 1, chi2limit = 0.05)
}
\arguments{
\item{matrix}{data.frame with numeric values}

\item{method}{switch to define which contingency value should be used:

"chi2" (default): test decision of the chi-square test for a defined decision niveau.
A significant relation of two variables/objects is marked with a numeric "1", a
negativ test result with a numeric "0".

"phi": Pearson's phi coefficient ("mean square contingency coefficient").

"cc": Pearson's contingency coefficient.

"lambda": Goodman and Kruskal's lambda value.}

\item{dim}{switch to define if the correlation matrix should be created
for columns or rows.

1 (default): table is created for column (variables) relations.

2: table is created for row (objects) relations.}

\item{chi2limit}{significance level for the test decision. Just relevant for method
"chi2".

default: 0.05 -> 5\%}
}
\value{
correlation matrix
}
\description{
\code{corrmat} returns a correlation matrix of a data.frame. Several different
correlation values can be choosen and the matrix can be created for column or row
relations.
}
\examples{
testmatrixrand <- data.frame(
   matrix(base::sample(0:1,400,replace=T), nrow=20, ncol=20)
)

corrmat(testmatrixrand, "chi2", chi2limit = 0.03)

corrmat(matrix = testmatrixrand, method = "lambda", dim = 2)

phicorrtable <- corrmat(matrix = testmatrixrand, method = "phi", dim = 1)
}

