% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lambdasym.R
\name{lambdasym}
\alias{lambdasym}
\title{Create a correlation matrix with Goodman and Kruskal's lambda value of an input data.frame}
\usage{
lambdasym(matrix, dim = 1)
}
\arguments{
\item{matrix}{data.frame with numeric values}

\item{dim}{switch to define, whether the correlation matrix should be created
for columns or rows.
1 (default): table is created for column (variables) relations.
2: table is created for row (objects) relations.}
}
\value{
lambda value correlation matrix
}
\description{
\code{lambdasym} returns a correlation matrix of a data.frame. The correlation value
is Goodman and Kruskal's lambda.
The correlation matrix can be created for column or row relations of the input
data.frame.
}
\examples{
testmatrixrand <- data.frame(
matrix(base::sample(0:1,400,replace=T), nrow=20, ncol=20)
)

# correlation table is created for the columns of the input data.frame testmatrixrand
lambdasym(testmatrixrand)
lambdasym(testmatrixrand, 1)
labmbdacolumns <- lambdasym(matrix = testmatrixrand, dim = 1)

# correlation table is created for the rows of the input data.frame testmatrixrand
lambdasym(testmatrixrand, 2)
lambdarows <- lambdasym(matrix = testmatrixrand, dim = 2)
}

