% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phicoeff.R
\name{phicoeff}
\alias{phicoeff}
\title{Create a phi coefficient correlation matrix of an input data.frame}
\usage{
phicoeff(matrix, dim = 1)
}
\arguments{
\item{matrix}{data.frame with numeric values}

\item{dim}{switch to define, whether the correlation matrix should be created
for columns or rows.
1 (default): table is created for column (variables) relations.
2: table is created for row (objects) relations.}
}
\value{
phi coefficient correlation matrix
}
\description{
\code{phicoeff} returns a correlation matrix of a data.frame. The correlation value
is the phi coefficient ("mean square contingency coefficient") by Karl Pearson.
The correlation matrix can be created for column or row relations of the input
data.frame.
}
\examples{
testmatrixrand <- data.frame(
matrix(base::sample(0:1,400,replace=T), nrow=20, ncol=20)
)

# correlation table is created for the columns of the input data.frame testmatrixrand
phicoeff(testmatrixrand)
phicoeff(testmatrixrand, 1)
phitabcolumns <- phicoeff(matrix = testmatrixrand, dim = 1)

# correlation table is created for the rows of the input data.frame testmatrixrand
phicoeff(testmatrixrand, 2)
phitabrows <- phicoeff(matrix = testmatrixrand, dim = 2)
}

