% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simplecc.R
\name{simplecc}
\alias{simplecc}
\title{Create a Pearson's contingency coefficient correlation matrix of an input data.frame}
\usage{
simplecc(matrix, dim = 1)
}
\arguments{
\item{matrix}{data.frame with numeric values}

\item{dim}{switch to define, whether the correlation matrix should be created
for columns or rows.
1 (default): table is created for column (variables) relations.
2: table is created for row (objects) relations.}
}
\value{
contingency coefficient correlation matrix
}
\description{
\code{simplecc} returns a correlation matrix of a data.frame. The correlation value
is Pearson's contingency coefficient.
The correlation matrix can be created for column or row relations of the input
data.frame.
}
\examples{
testmatrixrand <- data.frame(
matrix(base::sample(0:1,400,replace=T), nrow=20, ncol=20)
)

# correlation table is created for the columns of the input data.frame testmatrixrand
simplecc(testmatrixrand)
simplecc(testmatrixrand, 1)
cccolumns <- simplecc(matrix = testmatrixrand, dim = 1)

# correlation table is created for the rows of the input data.frame testmatrixrand
simplecc(testmatrixrand, 2)
ccrows <- simplecc(matrix = testmatrixrand, dim = 2)
}

